Experimental merge (portable_optimized kernels need manual updating)
# Conflicts:
#	tensorflow/compiler/mlir/lite/BUILD
#	tensorflow/compiler/mlir/lite/tests/prepare-tf.mlir
#	tensorflow/lite/kernels/internal/common.h
#	tensorflow/lite/micro/examples/person_detection/person_detection_test.cc
#	tensorflow/lite/micro/kernels/concatenation_test.cc
#	tensorflow/lite/micro/kernels/conv_test.cc
#	tensorflow/lite/micro/kernels/depthwise_conv_test.cc
#	tensorflow/lite/micro/kernels/elementwise_test.cc
#	tensorflow/lite/micro/kernels/logical_test.cc
#	tensorflow/lite/micro/kernels/maximum_minimum_test.cc
#	tensorflow/lite/micro/kernels/neg_test.cc
#	tensorflow/lite/micro/kernels/pack_test.cc
#	tensorflow/lite/micro/kernels/portable_optimized/conv.cc
#	tensorflow/lite/micro/kernels/portable_optimized/depthwise_conv.cc
#	tensorflow/lite/micro/kernels/portable_optimized/fully_connected.cc
#	tensorflow/lite/micro/kernels/prelu_test.cc
#	tensorflow/lite/micro/kernels/reshape_test.cc
#	tensorflow/lite/micro/kernels/split_test.cc
#	tensorflow/lite/micro/kernels/unpack_test.cc
#	tensorflow/lite/micro/memory_arena_threshold_test.cc
#	tensorflow/lite/micro/micro_error_reporter_test.cc
#	tensorflow/lite/micro/micro_interpreter_test.cc
#	tensorflow/lite/micro/test_helpers.cc
#	tensorflow/lite/micro/testing/memory_arena_threshold_test_supp.cc
#	tensorflow/lite/micro/testing/portable_optimized/memory_arena_threshold_test_supp.cc
#	tensorflow/lite/micro/testing/test_linux_binary.sh
#	tensorflow/lite/micro/testing/test_utils.h
#	tensorflow/lite/micro/tools/make/Makefile
#	tensorflow/lite/micro/tools/make/targets/windows_x86_makefile.inc

+++ tensorflow/lite/micro/testing/test_utils.h
@@ -72,9 +72,8 @@ private:
 // USE WITH CARE!! Returns pointer to data member of argument
 // so this object's lifetime must outlive any access to its underlying
 // data via this pointer.
-// Pass-by lvalue ref of argument protects against simply programmer
+// Pass-by lvalue ref of argument protects against simple programmer
 // oops but is by no means foolproof.
-
 inline TfLiteIntArray* IntArrayFromInitializer(
     std::initializer_list<int> &int_initializer) {
   return IntArrayFromInts(int_initializer.begin());
@@ -95,7 +94,8 @@ inline float MinFromZeroPointScale(const int zero_point, const float scale) {
 // Derives the quantization scaling factor from a min and max range.
 template <typename T>
 inline float ScaleFromMinMax(const float min, const float max) {
-  return (max - min) / ((std::numeric_limits<T>::max() * 1.0) -
+  return (max - min) /
+         static_cast<float>((std::numeric_limits<T>::max() * 1.0) -
                         std::numeric_limits<T>::min());
 }
